+++ src/game/savegame.cpp
@@ -159,7 +159,7 @@ int SaveGame(const std::string &filename)
 	// FIXME: probably not the right place for this
 	file.printf("GameCycle = %lu\n", GameCycle);
 	//Wyrmgus start
-	file.printf("GameTimeOfDay = %lu\n", GameTimeOfDay);
+	file.printf("GameTimeOfDay = %lu\n", (unsigned long) GameTimeOfDay);
 	//Wyrmgus end
 
 	file.printf("SetGodMode(%s)\n", GodMode ? "true" : "false");
